refactor(api): move sendVerificationCode method
from lib controller to src controller
