+++ api/lib/application/users/user-controller.js
@@ -2,7 +2,6 @@ import { usecases as devcompUsecases } from '../../../src/devcomp/domain/usecase
 import * as trainingSerializer from '../../../src/devcomp/infrastructure/serializers/jsonapi/training-serializer.js';
 import { evaluationUsecases } from '../../../src/evaluation/domain/usecases/index.js';
 import * as scorecardSerializer from '../../../src/evaluation/infrastructure/serializers/jsonapi/scorecard-serializer.js';
-import { usecases as IdentityAccessManagementUsecases } from '../../../src/identity-access-management/domain/usecases/index.js';
 import * as campaignParticipationSerializer from '../../../src/prescription/campaign-participation/infrastructure/serializers/jsonapi/campaign-participation-serializer.js';
 import * as userSerializer from '../../../src/shared/infrastructure/serializers/jsonapi/user-serializer.js';
 import * as requestResponseUtils from '../../../src/shared/infrastructure/utils/request-response-utils.js';
@@ -12,7 +11,6 @@ import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
 import * as campaignParticipationOverviewSerializer from '../../infrastructure/serializers/jsonapi/campaign-participation-overview-serializer.js';
 import * as certificationCenterMembershipSerializer from '../../infrastructure/serializers/jsonapi/certification-center-membership-serializer.js';
 import * as certificationEligibilitySerializer from '../../infrastructure/serializers/jsonapi/certification-eligibility-serializer.js';
-import * as emailVerificationSerializer from '../../infrastructure/serializers/jsonapi/email-verification-serializer.js';
 import * as participantResultSerializer from '../../infrastructure/serializers/jsonapi/participant-result-serializer.js';
 import * as profileSerializer from '../../infrastructure/serializers/jsonapi/profile-serializer.js';
 import * as sharedProfileForCampaignSerializer from '../../infrastructure/serializers/jsonapi/shared-profile-for-campaign-serializer.js';
@@ -245,20 +243,6 @@ const removeAuthenticationMethod = async function (request, h) {
   return h.response().code(204);
 };
 
-const sendVerificationCode = async function (
-  request,
-  h,
-  dependencies = { emailVerificationSerializer, requestResponseUtils },
-) {
-  const locale = dependencies.requestResponseUtils.extractLocaleFromRequest(request);
-  const i18n = request.i18n;
-  const userId = request.params.id;
-  const { newEmail, password } = await dependencies.emailVerificationSerializer.deserialize(request.payload);
-
-  await IdentityAccessManagementUsecases.sendVerificationCode({ i18n, locale, newEmail, password, userId });
-  return h.response().code(204);
-};
-
 const updateUserEmailWithValidation = async function (request, h, dependencies = { updateEmailSerializer }) {
   const userId = request.params.id;
   const code = request.payload.data.attributes.code;
@@ -348,7 +332,6 @@ const userController = {
   rememberUserHasSeenNewDashboardInfo,
   removeAuthenticationMethod,
   resetScorecard,
-  sendVerificationCode,
   updateUserDetailsForAdministration,
   updateUserEmailWithValidation,
 };
