+++ api/tests/unit/application/users/user-controller_test.js
@@ -8,10 +8,8 @@ import { usecases as devcompUsecases } from '../../../../src/devcomp/domain/usec
 import { evaluationUsecases } from '../../../../src/evaluation/domain/usecases/index.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { User } from '../../../../src/identity-access-management/domain/models/User.js';
-import { usecases as IdentityAccessManagementUsecases } from '../../../../src/identity-access-management/domain/usecases/index.js';
 import * as requestResponseUtils from '../../../../src/shared/infrastructure/utils/request-response-utils.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 describe('Unit | Controller | user-controller', function () {
   let userSerializer;
@@ -695,53 +693,6 @@ describe('Unit | Controller | user-controller', function () {
       expect(userAnonymizedDetailsForAdminSerializer.serialize).to.have.been.calledWithExactly(userDetailsForAdmin);
       expect(response.statusCode).to.equal(200);
       expect(response.source).to.deep.equal(anonymizedUserSerialized);
-    });
-  });
-
-  describe('#sendVerificationCode', function () {
-    it('should call the usecase to send verification code with code, email and locale', async function () {
-      // given
-      sinon.stub(IdentityAccessManagementUsecases, 'sendVerificationCode');
-      IdentityAccessManagementUsecases.sendVerificationCode.resolves();
-      const i18n = getI18n();
-      const userId = 1;
-      const locale = 'fr';
-      const newEmail = 'user@example.net';
-      const password = 'Password123';
-
-      const request = {
-        headers: { 'accept-language': locale },
-        i18n,
-        auth: {
-          credentials: {
-            userId,
-          },
-        },
-        params: {
-          id: userId,
-        },
-        payload: {
-          data: {
-            type: 'users',
-            attributes: {
-              newEmail,
-              password,
-            },
-          },
-        },
-      };
-
-      // when
-      await userController.sendVerificationCode(request, hFake);
-
-      // then
-      expect(IdentityAccessManagementUsecases.sendVerificationCode).to.have.been.calledWithExactly({
-        i18n,
-        locale,
-        newEmail,
-        password,
-        userId,
-      });
     });
   });
